backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic[email]==2.5.3

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9

# Security
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
cryptography==41.0.7
python-multipart==0.0.6

# Background tasks
celery==5.3.6
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Optional utilities
python-dateutil==2.8.2
